pyproject.toml
[tool.poetry]
name = "alignator"
version = "0.1.0"
description = ""
license = "GPL-3.0-or-later"
authors = ["Todoesverso <todoesverso@gmail.com>"]
packages = [{ include = "alignator" }]

[tool.poetry.scripts]
alignator = "alignator.bin.alignator:cli"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.4.1"}
#opencv-python = "^4.5.5"
# there is an import bug 
# https://github.com/opencv/opencv-python/issues/676
opencv-python = "4.5.5.64"
numpy = "^1.22.2"
scikit-image = "^0.19.1"
ffmpeg-python = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
black = "^22.1.0"
flake8-builtins = "^1.5.3"
flake8-blind-except = "^0.2.0"
flake8-logging-format = "^0.6.0"
flakeheaven = "^0.11.1"
memory-profiler = "^0.60.0"
pyproject-flake8 = "^0.0.1-alpha.4"
flake8-bugbear = "^22.3.23"

[tool.black]
line-length = 88
target-version = ['py36']

[tool.flakeheaven]
# optionally inherit from remote config (or local if you want)
base = "https://raw.githubusercontent.com/flakeheaven/flakeheaven/main/pyproject.toml"
# specify any flake8 options. For example, exclude "example.py":
exclude = ["example.py"]
# make output nice
format = "grouped"
# 80 chars aren't enough in 21 century
max_line_length = 90
# show line of source code in output
show_source = true

# list of plugins and rules for them
[tool.flakeheaven.plugins]
# include everything in pyflakes except F401
pyflakes = ["+*", "-F401"]
# enable only codes from S100 to S199
flake8-bandit = ["-*", "+S1??"]
# enable everything that starts from `flake8-`
"flake8-*" = ["+*"]
# explicitly disable plugin
flake8-docstrings = ["-*"]
mccabe = ["+*"]

[tool.flake8]
max-line-length = 88
max-complexity = 10
select = "C,E,F,W,B,B950"
ignore = "E203,E501,W503"
extend-ignore = "B008,B006"
exclude = ".git,__pycache__,*.egg-info,.nox,.pytest_cache,.mypy_cache"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
